# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Taxon Tag Service",
    "version": "1.0.0",
    "description": "Standalone tagging microservice — tag groups, entity tagging, audit workflow"
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer API Token"
      }
    },
...
  },
  "paths": {
...
    "/entities/{entityType}": {
      "get": {
        "tags": [
          "实体"
        ],
        "summary": "列出实体（支持标签过滤）",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "entityType",
            "in": "path"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 100,
              "default": 20
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "标签 ID 过滤（可多个）"
            },
            "required": false,
            "description": "标签 ID 过滤（可多个）",
            "name": "tagId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "按标签名称模糊过滤"
            },
            "required": false,
            "description": "按标签名称模糊过滤",
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "按 entityId 模糊搜索"
            },
            "required": false,
            "description": "按 entityId 模糊搜索",
            "name": "search",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "withTags",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "entityType": {
                                "type": "string"
                              },
                              "entityId": {
                                "type": "string"
                              },
                              "registeredAt": {
                                "anyOf": [
                                  {
                                    "type": "string",
                                    "format": "date-time"
                                  },
                                  {
                                    "type": "string"
                                  }
                                ]
                              },
                              "tags": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "id": {
                                      "type": "string"
                                    },
                                    "slug": {
                                      "type": "string"
                                    },
                                    "name": {
                                      "type": "string"
                                    },
                                    "groupId": {
                                      "type": "string"
                                    },
                                    "group": {
                                      "type": "object",
                                      "properties": {
                                        "id": {
                                          "type": "string"
                                        },
                                        "slug": {
                                          "type": "string"
                                        },
                                        "name": {
                                          "type": "string"
                                        }
                                      },
                                      "required": [
                                        "id",
                                        "slug",
                                        "name"
                                      ]
                                    },
                                    "source": {
                                      "type": "string"
                                    },
                                    "confidence": {
                                      "type": "number",
                                      "nullable": true
                                    },
                                    "status": {
                                      "type": "string"
                                    },
                                    "taggedAt": {
                                      "anyOf": [
                                        {
                                          "type": "string",
                                          "format": "date-time"
                                        },
                                        {
                                          "type": "string"
                                        }
                                      ]
                                    }
                                  },
                                  "required": [
                                    "id",
                                    "slug",
                                    "name",
                                    "groupId",
                                    "group",
                                    "source",
                                    "confidence",
                                    "status",
                                    "taggedAt"
                                  ]
                                }
                              }
                            },
                            "required": [
                              "entityType",
                              "entityId"
                            ]
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "items",
                        "total",
                        "page",
                        "pageSize"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/entities/{entityType}/{entityId}/tags": {
      "get": {
        "tags": [
          "实体标签"
        ],
        "summary": "获取实体标签列表",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "entityType",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "entityId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "pending",
                "rejected",
                "all"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "groupId": {
                            "type": "string"
                          },
                          "group": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "slug": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "slug",
                              "name"
                            ]
                          },
                          "source": {
                            "type": "string"
                          },
                          "confidence": {
                            "type": "number",
                            "nullable": true
                          },
                          "status": {
                            "type": "string"
                          },
                          "taggedAt": {
                            "anyOf": [
                              {
                                "type": "string",
                                "format": "date-time"
                              },
                              {
                                "type": "string"
                              }
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "slug",
                          "name",
                          "groupId",
                          "group",
                          "source",
                          "confidence",
                          "status",
                          "taggedAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "code",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "实体未注册",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "实体标签"
        ],
        "summary": "全量替换实体标签",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "entityType",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "entityId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tagIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "source": {
                    "type": "string",
                    "enum": [
                      "manual",
                      "ai",
                      "system",
                      "import"
                    ],
                    "default": "manual"
                  },
                  "confidence": {
                    "type": "number",
                    "nullable": true,
                    "minimum": 0,
                    "maximum": 1
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "pending",
                      "rejected"
                    ]
                  }
                },
                "required": [
                  "tagIds"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "标签不合法",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}